requirements.txt
dora-search
hydra-core
transformers
sentence-transformers
datasets
tabulate
pandas
pyyaml
tqdm
prettytable
colorama
matplotlib
conllu
torch_dct
nltk
plotext
torch
numpy
typing_extensions
entmax